10535 Activate Changes: Now able to reduce the number of maximum working apache processes
You can now configure how much apache processes are spawned during the activate changes phase.
In previous versions each site with a change always had one dedicated apache process.
All of these processes were spawned at the same time. This consumed quite an amount of memory when your WATO setup included lots (>50) of sites.

This werk introduces a scheduler mechanism, which ensures that
<ul>
<li>all sites with changes are updated</li>
<li>the number of apache processes do not go through the roof</li>
</ul>

Per default the optimal number is automatically calculated (up to 90% of RAM is used), which should be OK for most setups.
The new global settings option <tt>Maximum parallel processes during activate changes</tt> allows you to adjust this value.
The hardcoded minium value is 5.

+++ cmk/gui/watolib/activate_changes.py
@@ -30,6 +30,7 @@ import os
 import shutil
 import time
 import multiprocessing
+import psutil
 
 import cmk.utils
 import cmk.utils.daemon as daemon
@@ -66,9 +67,14 @@ from cmk.gui.watolib.changes import (
 )
 from cmk.gui.plugins.watolib import ConfigDomain
 
+import cmk.gui.gui_background_job as gui_background_job
+
+from cmk.gui.watolib.wato_background_job import WatoBackgroundJob
+
 # TODO: Make private
-PHASE_INITIALIZED = "initialized"  # Thread object has been initialized (not in thread yet)
-PHASE_STARTED = "started"  # Thread just started, nothing happened yet
+PHASE_INITIALIZED = "initialized"  # Process has been initialized (not in thread yet)
+PHASE_QUEUED = "queued"  # Process queued by site scheduler
+PHASE_STARTED = "started"  # Process just started, nothing happened yet
 PHASE_SYNC = "sync"  # About to sync
 PHASE_ACTIVATE = "activate"  # sync done activating changes
 PHASE_FINISHING = "finishing"  # Remote work done, finalizing local state
@@ -485,12 +491,17 @@ class ActivateChangesManager(ActivateChanges):
                 continue
 
             if site_state == {} or site_state["_phase"] == PHASE_INITIALIZED:
-                # Just been initialized. Treat as running as it has not been
-                # started and could not lock the site stat file yet.
+                # Just been initialized, not in queue yet
                 return True  # -> running
 
             # Check whether or not the process is still there
+            # The pid refers to
+            # - the site scheduler PID as long as the site is in queue
+            # - the site activate changes PID
+            # - None, if the site was locked by another process
             try:
+                if site_state["_pid"] is None:
+                    continue
                 os.kill(site_state["_pid"], 0)
                 return True  # -> running
             except OSError as e:
@@ -709,28 +720,12 @@ class ActivateChangesManager(ActivateChanges):
 
     def _start_activation(self):
         self._log_activation()
-        for site_id in self._sites:
-            self._start_site_activation(site_id)
-
-    def _start_site_activation(self, site_id):
-        self._log_site_activation(site_id)
-
-        # This is doing the first fork and the ActivateChangesSite() is doing the second
-        # (to avoid zombie processes when sync processes exit)
-        p = multiprocessing.Process(target=self._do_start_site_activation, args=[site_id])
-        p.start()
-        p.join()
-
-    def _do_start_site_activation(self, site_id):
-        try:
-            site_activation = ActivateChangesSite(site_id, self._activation_id,
-                                                  self._site_snapshot_file(site_id),
-                                                  self._prevent_activate)
-            site_activation.load()
-            site_activation.start()
-            os._exit(0)
-        except:
-            logger.exception()
+        site_snapshots = dict([
+            (site_id, self._site_snapshot_file(site_id)) for site_id in self._sites
+        ])
+        job = ActivateChangesSchedulerBackgroundJob(self._activation_id, site_snapshots,
+                                                    self._prevent_activate)
+        job.start()
 
     def _log_activation(self):
         log_msg = _("Starting activation (Sites: %s)") % ",".join(self._sites)
@@ -738,9 +733,6 @@ class ActivateChangesManager(ActivateChanges):
 
         if self._comment:
             log_audit(None, "activate-changes", "%s: %s" % (_("Comment"), self._comment))
-
-    def _log_site_activation(self, site_id):
-        log_audit(None, "activate-changes", _("Started activation of site %s") % site_id)
 
     def get_state(self):
         state = {
@@ -805,10 +797,101 @@ class ActivateChangesManager(ActivateChanges):
         return ids
 
 
+@gui_background_job.job_registry.register
+class ActivateChangesSchedulerBackgroundJob(WatoBackgroundJob):
+    job_prefix = "activate-changes-scheduler"
+    housekeeping_max_age_sec = 86400 * 30
+    housekeeping_max_count = 10
+
+    @classmethod
+    def gui_title(cls):
+        return _("Activate Changes Scheduler")
+
+    def __init__(self, activation_id, site_snapshots, prevent_activate):
+        super(ActivateChangesSchedulerBackgroundJob,
+              self).__init__("%s-%s" % (self.job_prefix, activation_id),
+                             deletable=False,
+                             stoppable=False)
+        self._activation_id = activation_id
+        self._site_snapshots = site_snapshots
+        self._prevent_activate = prevent_activate
+        self._job_interface = None
+        self.set_function(self._schedule_sites)
+
+    def _send_progress_update(self, info):
+        self._job_interface.send_progress_update("%s %s" % (render.time_of_day(time.time()), info))
+
+    def _schedule_sites(self, job_interface):
+        self._job_interface = job_interface
+        # Prepare queued jobs
+
+        self._send_progress_update(_("Activate Changes Scheduler started"))
+        queued_jobs = self._get_queued_jobs()
+
+        self._send_progress_update(_("Going to update %d sites") % len(queued_jobs))
+
+        running_jobs = []
+        max_jobs = self._get_maximum_concurrent_jobs()
+        while queued_jobs or len(running_jobs) > 0:
+            # Housekeeping, remove finished jobs
+            for job in running_jobs[:]:
+                if job.is_alive():
+                    continue
+                self._send_progress_update(_("Finished site update: %s") % job.site_id)
+                job.join()
+                running_jobs.remove(job)
+
+            time.sleep(0.1)
+
+            # Continue if at max concurrent jobs
+            if len(running_jobs) == max_jobs:
+                continue
+
+            # Start new jobs
+            while queued_jobs:
+                job = queued_jobs.pop(0)
+                self._send_progress_update(_("Starting site update: %s") % job.site_id)
+                job.start()
+                running_jobs.append(job)
+                if len(running_jobs) == max_jobs:
+                    break
+
+        self._job_interface.send_result_message(_("Activate changes finished"))
+
+    def _get_maximum_concurrent_jobs(self):
+        if config.wato_activate_changes_concurrency == "auto":
+            processes = self._max_processes_based_on_ram()
+        else:  # (maximum, 23)
+            processes = config.wato_activate_changes_concurrency[1]
+        return max(5, processes)
+
+    def _max_processes_based_on_ram(self):
+        # This process will be forked mulitple times
+        # Determine its current rss usage and compute a reasonable maximum value
+        try:
+            # We are going to fork this process
+            process = psutil.Process(os.getpid())
+            size = process.memory_info().rss
+            return (0.9 * psutil.virtual_memory().available) // size
+        except RequestTimeout:
+            raise
+        except:
+            return 1
+
+    def _get_queued_jobs(self):
+        queued_jobs = []
+        for site_id in sorted(self._site_snapshots):
+            site_job = ActivateChangesSite(site_id, self._activation_id,
+                                           self._site_snapshots[site_id], self._prevent_activate)
+            site_job.load()
+            if site_job.lock_activation():
+                queued_jobs.append(site_job)
+        return queued_jobs
+
+
 class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
     def __init__(self, site_id, activation_id, site_snapshot_file, prevent_activate=False):
         super(ActivateChangesSite, self).__init__()
-
         self._site_id = site_id
         self._site_changes = []
         self._activation_id = activation_id
@@ -828,6 +911,10 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
         self._expected_duration = 10.0
 
         self._set_result(PHASE_INITIALIZED, _("Initialized"))
+
+    @property
+    def site_id(self):
+        return self._site_id
 
     def load(self):
         super(ActivateChangesSite, self).load()
@@ -880,7 +967,13 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
     def _do_run(self):
         try:
             self._time_started = time.time()
-            self._lock_activation()
+
+            # Update PID
+            # Initially the SiteScheduler set its own PID into the sites state file
+            # The PID itself is used to detect whether the sites activation process is still running
+            self._mark_running()
+
+            log_audit(None, "activate-changes", _("Started activation of site %s") % self._site_id)
 
             if self.is_sync_needed(self._site_id):
                 self._synchronize_site()
@@ -935,33 +1028,40 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
         html_code += "</div>"
         return html_code
 
-    def _lock_activation(self):
+    def lock_activation(self):
         # This locks the site specific replication status file
         repl_status = _load_site_replication_status(self._site_id, lock=True)
+        got_lock = False
         try:
             if self._is_currently_activating(repl_status):
-                raise MKGeneralException(
+                self._set_result(
+                    PHASE_DONE,
+                    _("Locked"),
+                    status_details=
                     _("The site is currently locked by another activation process. Please try again later"
-                     ))
+                     ),
+                    state=STATE_WARNING)
+                return False
 
-            # This is needed to detect stale activation progress entries
-            # (where activation thread is not running anymore)
-            self._mark_running()
+            got_lock = True
+            self._mark_queued()
+            return True
         finally:
             # This call unlocks the replication status file after setting "current_activation"
             # which will prevent other users from starting an activation for this site.
-            _update_replication_status(self._site_id, {"current_activation": self._activation_id})
+            # If the site was already locked, simply release the lock without further changes
+            if got_lock:
+                _update_replication_status(self._site_id,
+                                           {"current_activation": self._activation_id})
+            else:
+                store.release_lock(_site_replication_status_path(self._site_id))
 
-    def _is_currently_activating(self, rep_status):
-        if not rep_status.get("current_activation"):
+    def _is_currently_activating(self, site_rep_status):
+        current_activation_id = site_rep_status.get("current_activation")
+        if not current_activation_id:
             return False
 
-        #
         # Is this activation still in progress?
-        #
-
-        current_activation_id = rep_status.get(self._site_id, {}).get("current_activation")
-
         manager = ActivateChangesManager()
         manager.load()
 
@@ -975,7 +1075,13 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
 
         return False
 
+    def _mark_queued(self):
+        # Is set by site scheduler
+        self._pid = os.getpid()
+        self._set_result(PHASE_QUEUED, _("Queued"))
+
     def _mark_running(self):
+        # Is set by active site process
         self._pid = os.getpid()
         self._set_result(PHASE_STARTED, _("Started"))
 
@@ -1113,9 +1219,15 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
         self._save_state()
 
     def _set_status_details(self, phase, status_details):
-        self._status_details = _("Started at: %s.") % render.time_of_day(self._time_started)
+        # As long as the site is lying in queue, there is no time started
+        if phase == PHASE_QUEUED:
+            self._status_details = _("Queued for update")
+        else:
+            self._status_details = _("Started at: %s.") % render.time_of_day(self._time_started)
 
-        if phase != PHASE_DONE:
+        if phase == PHASE_DONE:
+            self._status_details += _(" Finished at: %s.") % render.time_of_day(self._time_ended)
+        elif phase != PHASE_QUEUED:
             estimated_time_left = self._expected_duration - (time.time() - self._time_started)
             if estimated_time_left < 0:
                 self._status_details += " " + _("Takes %.1f seconds longer than expected") % \
@@ -1123,8 +1235,6 @@ class ActivateChangesSite(multiprocessing.Process, ActivateChanges):
             else:
                 self._status_details += " " + _("Approximately finishes in %.1f seconds") % \
                                                                         estimated_time_left
-        else:
-            self._status_details += _(" Finished at: %s.") % render.time_of_day(self._time_ended)
 
         if status_details:
             self._status_details += "<br>%s" % status_details
